controladores y migraciones

+++ app/Http/Controllers/ClientesController.php
@@ -43,31 +43,52 @@ class ClientesController extends Controller
      */
     public function store(ClientesRequest $request)
     {
-        $nombre_qr = $request->rut.'-'.$request->verificador;
-        $qr_code = QRCode::text($nombre_qr)
-        ->setOutfile('./img/qr-code/'.$nombre_qr.'.png')
-        ->setSize(8)
-        ->setMargin(2)
-        ->png();
+        if(\Auth::user()->tipo_usuario == 'Admin'){
+            $nombre_qr = $request->rut.'-'.$request->verificador;
+            $clave = $request->rut.''.$request->verificador;
+            //dd($clave);
 
-        $url_img = "img/qr-code/".$nombre_qr.".png";
-        $qr = new CodigoQr();
-        $qr->codigo=$url_img;
-        $qr->codigo_recupera=1234;
-        $qr->status="Activo";
-        $qr->save();
-        
-        $clientes = new Clientes();
-        $clientes->id_qr=$qr->id;
-        $clientes->nombres=$request->nombres;
-        $clientes->apellidos=$request->apellidos;
-        $clientes->rut = $request->rut.'-'.$request->verificador;
-        $clientes->email=$request->email;
-        $clientes->status=$request->status;
-        $clientes->save();
+            $qr_code = QRCode::text($nombre_qr)
+            ->setOutfile('./img/qr-code/'.$nombre_qr.'.png')
+            ->setSize(8)
+            ->setMargin(2)
+            ->png();
 
-        toastr()->success('Éxito!!', ' Cliente registrado satisfactoriamente');
-        return redirect()->to('clientes');
+            $url_img = "img/qr-code/".$nombre_qr.".png";
+            $qr = new CodigoQr();
+            $qr->codigo=$url_img;
+            $qr->codigo_recupera=1234;
+            $qr->status="Activo";
+            $qr->save();
+
+            $usuario = new User();
+            $usuario->usuario=$request->usuario;
+            $usuario->email=$request->email;
+            $nueva_clave=\Hash::make($clave);
+            $usuario->password=$nueva_clave;
+            $usuario->tipo_usuario="Cliente";
+            $usuario->save();
+            
+            $clientes = new Clientes();
+            $clientes->id_qr=$qr->id;
+            $clientes->id_usuario=$usuario->id;
+            $clientes->nombres=$request->nombres;
+            $clientes->apellidos=$request->apellidos;
+            $clientes->rut = $request->rut.'-'.$request->verificador;
+            $clientes->status=$request->status;
+            $clientes->save();
+
+            toastr()->success('Éxito!!', ' Cliente registrado satisfactoriamente');
+            return redirect()->to('clientes');
+        }else{
+            toastr()->warning('no puede acceder!!', 'ACCESO DENEGADO');
+            return redirect()->back();
+        }
+    }
+
+    public function carnet_qr(){
+        $pdf = \PDF::loadView('pdf/carnet_qr');
+        return $pdf->stream('carnet_qr.pdf');
     }
 
     /**
@@ -106,39 +127,49 @@ class ClientesController extends Controller
 
     public function editar(Request $request)
     {
-        $clientes = Clientes::find($id);
-        $clientes->nombres=$request->nombres;
-        $clientes->apellidos=$request->apellidos;
-        $clientes->rut = $request->rut.'-'.$request->verificador;
-        $clientes->email=$request->email;
-        $clientes->status=$request->status;
-        $clientes->save();
+        if(\Auth::user()->tipo_usuario == 'Admin'){
+            $clientes = Clientes::find($id);
+            $clientes->nombres=$request->nombres;
+            $clientes->apellidos=$request->apellidos;
+            $clientes->rut = $request->rut.'-'.$request->verificador;
+            $clientes->email=$request->email;
+            $clientes->status=$request->status;
+            $clientes->save();
 
-        toastr()->success('Éxito!!', ' Cliente registrado satisfactoriamente');
-        return redirect()->to('clientes');
+            toastr()->success('Éxito!!', ' Cliente registrado satisfactoriamente');
+            return redirect()->to('clientes');
+        }else{
+            toastr()->warning('no puede acceder!!', 'ACCESO DENEGADO');
+            return redirect()->back();
+        }
     }
 
     public function eliminar(Request $request)
     {
-        //dd($request->all());
-        $password = $request['password'];
-        $consulta = User::where('id',\Auth::User()->id)->first();
-        $hashedPassword = $consulta->password;
+        if(\Auth::user()->tipo_usuario == 'Admin'){
+            //dd($request->all());
+            $password = $request['password'];
+            $consulta = User::where('id',\Auth::User()->id)->first();
+            $hashedPassword = $consulta->password;
 
-        if (\Hash::check($password, $hashedPassword)) {
-            DB::table('clientes')->delete($request->id);
-            $data=CodigoQr::where('id',$request->id_qr)->first();
-            $image_path = $data->codigo;  // the value is : localhost/project/image/filename.format
+            if (\Hash::check($password, $hashedPassword)) {
+                DB::table('clientes')->delete($request->id);
+                $data=CodigoQr::where('id',$request->id_qr)->first();
+                $image_path = $data->codigo;  // the value is : localhost/project/image/filename.format
 
-            if(file_exists($image_path)) {
-                unlink($image_path);
+                if(file_exists($image_path)) {
+                    unlink($image_path);
+                }
+                DB::table('codigo_qr')->delete($request->id_qr);
+                toastr()->success('Éxito!!', ' Cliente eliminado satisfactoriamente');
+                return redirect()->to('clientes');
+            }else{
+                toastr()->error('Error!!', ' Contraseña incorrecta');
+                return redirect()->to('clientes');
             }
-            DB::table('codigo_qr')->delete($request->id_qr);
-            toastr()->success('Éxito!!', ' Cliente eliminado satisfactoriamente');
-            return redirect()->to('clientes');
         }else{
-            toastr()->error('Error!!', ' Contraseña incorrecta');
-            return redirect()->to('clientes');
+            toastr()->warning('no puede acceder!!', 'ACCESO DENEGADO');
+            return redirect()->back();
         }
     }
